(MAINT) Reduce and Improve Knack Table
It is unlikely most players will have access to a d30 to generate a
knack at character creation. This change cuts the list of possible
combinations down to 20 for each category, enabling a tighter list
of knacks whose combinations make more sense. This still leaves us
with over a million possible combinations and (assuming only 10% are
valid), over a hundred thousand possible knacks.